Refactor - create connect wrapper that will be used for proper typing - Once we have State type, we will be able to supply precise typing information
