+++ app/frontend/components/app.tsx
@@ -1,5 +1,5 @@
 import {h} from 'preact'
-import {connect} from '../libs/unistore/preact'
+import {connect} from '../helpers/connect'
 
 import {TopLevelRouter} from './router'
 import Welcome from './common/welcome'
