+++ app/frontend/components/autoLogout.ts
@@ -1,5 +1,5 @@
 import {Component} from 'preact'
-import {connect} from '../libs/unistore/preact'
+import {connect} from '../helpers/connect'
 import actions from '../actions'
 import {ADALITE_CONFIG} from '../config'
