+++ app/frontend/components/common/contactForm.tsx
@@ -1,5 +1,5 @@
 import {h, Component} from 'preact'
-import {connect} from '../../libs/unistore/preact'
+import {connect} from '../../helpers/connect'
 import actions from '../../actions'
 import Modal from './modal'
 import {ADALITE_CONFIG} from '../../config'
@@ -122,6 +122,6 @@ class ContactForm extends Component<Props, {submitted: boolean}> {
 }
 
 export default connect(
-  {},
+  null,
   actions
 )(ContactForm)
