+++ app/frontend/components/common/footer.tsx
@@ -1,5 +1,5 @@
 import {h} from 'preact'
-import {connect} from '../../libs/unistore/preact'
+import {connect} from '../../helpers/connect'
 import actions from '../../actions'
 import Branding from './branding'
 import {
