+++ app/frontend/components/common/navbar/navbarAuth.tsx
@@ -1,5 +1,5 @@
 import {h, Component} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import {ADALITE_CONFIG} from '../../../config'
 const APP_VERSION = ADALITE_CONFIG.ADALITE_APP_VERSION
