+++ app/frontend/components/common/navbar/navbarUnauth.tsx
@@ -1,5 +1,5 @@
 import {h} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import {ADALITE_CONFIG} from '../../../config'
 import isLeftClick from '../../../helpers/isLeftClick'
