+++ app/frontend/components/common/welcome.tsx
@@ -1,5 +1,5 @@
 import {h, Component} from 'preact'
-import {connect} from '../../libs/unistore/preact'
+import {connect} from '../../helpers/connect'
 import actions from '../../actions'
 import Modal from './modal'
 import Alert from './alert'
@@ -115,6 +115,6 @@ class Welcome extends Component<WelcomeProps, {dontShowAgainCheckbox: boolean}> 
 }
 
 export default connect(
-  {},
+  null,
   actions
 )(Welcome)
