+++ app/frontend/components/pages/login/logoutNotification.tsx
@@ -1,5 +1,5 @@
 import {h} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import Modal from '../../common/modal'
 
@@ -38,6 +38,6 @@ class LogoutNotification {
 }
 
 export default connect(
-  {},
+  null,
   actions
 )(LogoutNotification)
