+++ app/frontend/components/pages/receiveAda/myAddresses.tsx
@@ -1,8 +1,8 @@
 import {h} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import AddressItem from './addressItem'
 
-const MyAddresses = connect('ownAddressesWithMeta')(({ownAddressesWithMeta}) => (
+const _MyAddresses = ({ownAddressesWithMeta}) => (
   <div className="addresses card">
     <h2 className="card-title">My Addresses</h2>
     <div className="addresses-content">
@@ -11,6 +11,9 @@ const MyAddresses = connect('ownAddressesWithMeta')(({ownAddressesWithMeta}) => 
       ))}
     </div>
   </div>
-))
+)
+
+
+const MyAddresses = connect((state) => ({ownAddressesWithMeta: state.ownAddressesWithMeta}))(_MyAddresses)
 
 export default MyAddresses
