+++ app/frontend/components/pages/sendAda/customInputButton.tsx
@@ -1,5 +1,5 @@
 import {h} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import {AdaIcon} from '../../common/svg'
 import {toAda} from '../../../helpers/adaConverters'
