+++ app/frontend/components/pages/sendAda/donationButtons.tsx
@@ -1,5 +1,5 @@
 import {h, Component} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 import {AdaIcon} from '../../common/svg'
 import {ADALITE_CONFIG} from '../../../config'
