+++ app/frontend/components/pages/sendAda/sendAdaPage.tsx
@@ -1,5 +1,5 @@
 import {h, Component} from 'preact'
-import {connect} from '../../../libs/unistore/preact'
+import {connect} from '../../../helpers/connect'
 import actions from '../../../actions'
 
 import {getTranslation} from '../../../translations'
