+++ app/frontend/helpers/connect.ts
@@ -1,0 +1,3 @@
+import {connect} from '../libs/unistore/preact'
+
+export {connect}
